// projects/ngx-epic-fhir-client/src/lib/model/bundleEndpointDSTU2.ts
/**
 * Epic on Fhir
 * All of the components of the Change Healthcare Interoperability Apis
 *
 * OpenAPI spec version: 1.0.0
 * Contact: bmain@lumeris.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { BundleEntryEndpointDSTU2 } from './bundleEntryEndpointDSTU2';

/**
 * Contains a collection of resources.
 */
export interface BundleEndpointDSTU2 {
  /**
   * Entry in the bundle.
   */
  entry?: Array<BundleEntryEndpointDSTU2>;
}